requirements.txt
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.4
djangorestframework==3.17.1
docstring_parser==0.18.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
packaging==26.1
psycopg2-binary==2.9.11
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
whitenoise==6.12.0
